Updated Npcs Maps Items and Animations (Quests too?) To Integers
So far seems to have no crash or side effect, maps created and tested through multiple client and server restarts without side effect, npcs and items seem to also have no immediate side effects so im leaving it here unless bugs are found at which ill correct immediately :D

+++ Source/Client/Source/Modules/Database/modQuest.vb
@@ -3,7 +3,7 @@
 Friend Module ModQuest
 #Region "Global Info"
     'Constants
-    Friend Const MaxQuests As Byte = 250
+    Friend Const MaxQuests As Integer = 250
     'Friend Const MAX_TASKS As Byte = 10
     'Friend Const MAX_REQUIREMENTS As Byte = 10
     Friend Const EditorTasks As Byte = 7
